[REFACTOR] 마이페이지 스타일 정리
