+++ src/components/mypage/article/neighbors/MyNeighbors.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import DropDown from "../../DropDown";
+import DropDown from "../../elements/DropDown";
 
 import NeighborCard from "./elements/NeighborCard";
